requirements.txt
python-dotenv
requests
pydub
spotipy
openai
tiktoken
youtube_dl
